+++ actiontext/test/dummy/app/models/validated_message.rb
@@ -1,0 +1,3 @@
+class ValidatedMessage < Message
+  validates :body, presence: true
+end