+++ actionview/lib/action_view/helpers/tags/date_select.rb
@@ -13,6 +13,8 @@ module ActionView
         end
 
         def render
+          @validator.validate!(@html_options)
+
           error_wrapping(datetime_selector(@options, @html_options).public_send("select_#{select_type}").html_safe)
         end