+++ actionview/lib/action_view/helpers/tags/text_field.rb
@@ -13,6 +13,7 @@ module ActionView
           options["size"] = options["maxlength"] unless options.key?("size")
           options["type"] ||= field_type
           options["value"] = options.fetch("value") { value_before_type_cast } unless field_type == "file"
+          @validator.validate!(options)
           add_default_name_and_id(options)
           tag("input", options)
         end